yelp_backend_v2/owner_service/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
bcrypt
python-multipart
pillow
python-dotenv
pydantic[email]
cryptography
pymongo
motor
